+++ allianceutils/__init__.py
@@ -1,6 +1,9 @@
 import unittest
 
-default_app_config = 'allianceutils.apps.AllianceUtilsAppConfig'
+import django
+
+if django.VERSION < (3, 2):
+  default_app_config = 'allianceutils.apps.AllianceUtilsAppConfig'
 
 # All the tests are in test_allianceutils
 # Django's test autodiscovery will try to recursively import everything which causes asynctask to fail if we're running
@@ -9,4 +12,4 @@ def load_tests(*args, **kwargs):
   empty_suite = unittest.TestSuite()
   return empty_suite
 
-__version__ = "1.2.0"
+__version__ = "2.0.0.dev0"