+++ allianceutils/checks.py
@@ -3,31 +3,28 @@ from collections import defaultdict
 import inspect
 from pathlib import Path
 import re
-from typing import Callable
 from typing import Dict
 from typing import Iterable
 from typing import List
 from typing import Mapping
 from typing import Optional
 from typing import Type
+from typing import Union
 
-import django
 from django.apps import apps
 from django.apps.config import AppConfig
 from django.conf import settings
+from django.core.checks import CheckMessage
 from django.core.checks import Error
+from django.core.checks import Info
 from django.core.checks import Warning
 from django.db.models import Model
 from django.urls import get_resolver
+from django.urls import URLResolver
 
 from allianceutils.util import camel_to_underscore
 from allianceutils.util import get_firstparty_apps
 from allianceutils.util import underscore_to_camel
-
-if django.VERSION >= (2, 0):
-    from django.urls import URLResolver
-else:
-    from django.conf.urls import RegexURLResolver as URLResolver
 
 # W001 not used
 # W002 not used
@@ -43,12 +40,19 @@ ID_WARNING_GIT_HOOKS = 'allianceutils.W008'
 ID_ERROR_GIT_HOOKS = 'allianceutils.E008'
 ID_ERROR_EXPLICIT_TABLE_NAME = 'allianceutils.E009'
 ID_ERROR_EXPLICIT_TABLE_NAME_LOWERCASE = 'allianceutils.E010'
+ID_INFO_EXPLICIT_TABLE_NAME_LOWERCASE = 'allianceutils.I010'
 ID_ERROR_FIELD_NAME_NOT_CAMEL_FRIENDLY = 'allianceutils.E011'
+
+try:
+    Pattern = re.Pattern
+except AttributeError:
+    # Until python 3.7 re.Pattern was not exposed (was only present internally as _pattern_type)
+    Pattern = object
 
 
 def find_candidate_models(
-        app_configs: Optional[Iterable[AppConfig]],
-        ignore_labels: Iterable[str] = None
+    app_configs: Optional[Iterable[AppConfig]],
+    ignore_labels: Iterable[Union[str, Pattern]] = None
 ) -> Dict[str, Type[Model]]:
     """
     Given a list of labels to ignore, return models whose app_label or label is NOT in ignore_labels.
@@ -60,7 +64,14 @@ def find_candidate_models(
     if ignore_labels is None:
         ignore_labels = []
 
-    ignore_labels = set(ignore_labels)
+    def should_ignore(label: str) -> bool:
+        return (
+            # string match
+            any(s == label for s in ignore_labels) or
+            # regex pattern match
+            any(p.match(label) for p in ignore_labels if hasattr(p, "match"))
+        )
+
     candidate_models: Dict[str, Type[Model]] = {}
 
     for app_config in app_configs:
@@ -68,7 +79,7 @@ def find_candidate_models(
             model._meta.label: model
             for model
             in app_config.get_models()
-            if model._meta.app_label not in ignore_labels and model._meta.label not in ignore_labels
+            if not should_ignore(model._meta.app_label) and not should_ignore(model._meta.label)
         })
 
     return candidate_models
@@ -78,11 +89,11 @@ class CheckUrlTrailingSlash:
     expect_trailing_slash: bool
     ignore_attrs: Mapping[str, Iterable[str]]
 
-    def __init__(self, expect_trailing_slash: bool, ignore_attrs: Mapping[str, Iterable[str]]={}):
+    def __init__(self, expect_trailing_slash: bool, ignore_attrs: Mapping[str, Iterable[str]] = {}):
         self.expect_trailing_slash = expect_trailing_slash
         self.ignore_attrs = ignore_attrs
 
-    def __call__(self, app_configs: Iterable[AppConfig], **kwargs):
+    def __call__(self, app_configs: Iterable[AppConfig], **kwargs) -> List[CheckMessage]:
         # We ignore app_configs; so does django core check_url_settings()
         # Consider where ROOT_URLCONF points app A urlpatterns which include() app B urlpatterns
         # which define a URL to view in app C -- which app was the one that owned the URL?
@@ -100,8 +111,8 @@ class CheckUrlTrailingSlash:
         }
         _ignore_attrs.update(self.ignore_attrs)
 
-        def check_resolver(resolver: URLResolver, depth: int=0):
-            warnings = []
+        def check_resolver(resolver: URLResolver, depth: int = 0) -> List[CheckMessage]:
+            messages = []
             for url_pattern in resolver.url_patterns:
 
                 # look
@@ -126,7 +137,7 @@ class CheckUrlTrailingSlash:
 
                 if hasattr(url_pattern, 'url_patterns'):
                     # is a resolver, not a pattern: recurse
-                    warnings.extend(check_resolver(url_pattern, depth+1))
+                    messages.extend(check_resolver(url_pattern, depth+1))
 
                 elif regex_pattern.endswith('/$') != self.expect_trailing_slash:
                     try:
@@ -136,7 +147,7 @@ class CheckUrlTrailingSlash:
                         # django <2.0 regex urls
                         description = url_pattern.describe()
 
-                    warnings.append(
+                    messages.append(
                         Warning(
                             f'The URL pattern {description} is inconsistent with expect_trailing_slash',
                             obj=url_pattern,
@@ -144,12 +155,12 @@ class CheckUrlTrailingSlash:
                         )
                     )
 
-            return warnings
+            return messages
 
         return check_resolver(get_resolver())
 
 
-def check_git_hooks(app_configs: Iterable[AppConfig], **kwargs):
+def check_git_hooks(app_configs: Iterable[AppConfig], **kwargs) -> List[CheckMessage]:
     git_path = Path(settings.PROJECT_DIR, '.git')
 
     # handle the case where .git is a file rather than a directory
@@ -159,12 +170,12 @@ def check_git_hooks(app_configs: Iterable[AppConfig], **kwargs):
     except OSError:
         pass
 
-    warnings = []
+    messages = []
 
     if not git_path.exists():
         if settings.DEBUG:
             # If in dev mode then there should be a .git dir
-            warnings.append(
+            messages.append(
                 Warning(
                     "DEBUG is true but can't find a .git dir; are you trying to use DEBUG in production?",
                     obj=git_path,
@@ -192,7 +203,7 @@ def check_git_hooks(app_configs: Iterable[AppConfig], **kwargs):
 
         if not found_installed_hooks:
             (warning_type, warning_id) = (Error, ID_ERROR_GIT_HOOKS) if settings.DEBUG else (Warning, ID_WARNING_GIT_HOOKS)
-            warnings.append(
+            messages.append(
                 warning_type(
                     "git hooks are not configured (husky should be installed or .git/hooks should be a symlink to the git-hooks directory)",
                     obj=git_hooks_path,
@@ -200,24 +211,24 @@ def check_git_hooks(app_configs: Iterable[AppConfig], **kwargs):
                 )
             )
 
-    return warnings
+    return messages
 
 
-def check_admins(app_configs: Iterable[AppConfig], **kwargs):
-    errors = []
+def check_admins(app_configs: Iterable[AppConfig], **kwargs) -> List[CheckMessage]:
+    messages = []
     if not settings.AUTOMATED_TESTS and not settings.DEBUG:
         if len(settings.ADMINS) == 0:
-            errors.append(
+            messages.append(
                 Error(
                     "settings.ADMINS should not be empty",
                     obj='settings',
                     id=ID_ERROR_ADMINS,
                 )
             )
-    return errors
+    return messages
 
 
-def check_db_constraints(app_configs: Iterable[AppConfig], **kwargs):
+def check_db_constraints(app_configs: Iterable[AppConfig], **kwargs) -> List[CheckMessage]:
     """
     If using django-db-constraints, constraint identifiers can be supplied
     that are longer than the max identifier length for Postgres
@@ -238,46 +249,73 @@ def check_db_constraints(app_configs: Iterable[AppConfig], **kwargs):
     known_constraints = defaultdict(list)
     for app_config in app_configs.values():
         for model in app_config.get_models():
+            # native django constraints
+            if hasattr(model._meta, 'constraints'):
+                for constraint in model._meta.constraints:
+                    known_constraints[_truncate_constraint_name(constraint.name)].append((model, constraint.name))
+            # constraints from the django-db-constraints package
             if hasattr(model._meta, 'db_constraints'):
                 for constraint_name in model._meta.db_constraints.keys():
                     known_constraints[_truncate_constraint_name(constraint_name)].append((model, constraint_name))
 
-    errors = []
+    messages = []
     for truncated_name, model_constraints in known_constraints.items():
         if len(model_constraints) == 1:
             continue
-        _models = ['%s.%s' % (model._meta.app_label, model.__name__) for model, _ in model_constraints]
+        _models = [f'{model._meta.app_label}.{model.__name__}' for model, _ in model_constraints]
         models_string = ', '.join(_models)
         for _model, constraint_name in model_constraints:
-            errors.append(
+            messages.append(
                 Error(
-                    '%s constraint %s is not unique' % (_model._meta.label, constraint_name),
-                    hint='Constraint truncates to %s' % truncated_name.decode('utf-8', 'replace'),
+                    f'{_model._meta.label} constraint {constraint_name} is not unique',
+                    hint='Constraint truncates to ' + truncated_name.decode('utf-8', 'replace'),
                     obj=models_string,
                     id=ID_ERROR_DB_CONSTRAINTS,
                 )
             )
 
-    return errors
+    return messages
 
-def _check_explicit_table_names_on_model(model: Type[Model], enforce_lowercase: bool) -> Iterable[Type[Error]]:
+
+def _check_explicit_table_names_on_model(model: Type[Model], enforce_lowercase: bool) -> List[CheckMessage]:
     """
     Use an ast to check if a model has the db_table meta option set.
     This is done this way because a model instance's db_table is always
     populated even if with that of the default.
     """
-    errors = []
+    messages = []
     found = None
-    tree = ast.parse(inspect.getsource(model))
+    try:
+        source = inspect.getsource(model)
+    except OSError:
+        # if a model class is dynamically created then we can't inspect the source code
+        # (eg this happens with audit models)
+        message = f"Can't get source for {model.__name__}"
+        # warnings.warn(message)
+        messages.append(Info(
+            message,
+            hint="Dynamically generated model source code can't be introspected",
+            obj=model,
+            id=ID_INFO_EXPLICIT_TABLE_NAME_LOWERCASE,
+        ))
+        return messages
+    tree = ast.parse(source)
     for node in ast.walk(tree):
         if isinstance(node, ast.ClassDef) and node.name == 'Meta':
             for sub_node in node.body:
                 if isinstance(sub_node, ast.Assign):
                     if sub_node.targets[0].id == 'db_table':
-                        found = sub_node.value.s
+                        # in python <=3.7 it will be an ast.Str
+                        # in python >=3.8 it will be an ast.Constant
+                        if isinstance(sub_node.value, (ast.Constant, ast.Str)):
+                            found = sub_node.value.s
+                        else:
+                            # If it's an expression then we don't know what it's going to evaluate it
+                            # so we're just going to assume it's ok
+                            found = True
                         break
     if not found:
-        errors.append(
+        messages.append(
             Error(
                 'Explicit table name required',
                 hint=f'Add db_table setting to {model._meta.label} model Meta',
@@ -285,8 +323,8 @@ def _check_explicit_table_names_on_model(model: Type[Model], enforce_lowercase: 
                 id=ID_ERROR_EXPLICIT_TABLE_NAME,
             )
         )
-    elif enforce_lowercase and not found.islower():
-        errors.append(
+    elif enforce_lowercase and hasattr(found, "islower") and not found.islower():
+        messages.append(
             Error(
                 'Table names must be lowercase',
                 hint=f'Check db_table setting for {model._meta.label}',
@@ -295,7 +333,7 @@ def _check_explicit_table_names_on_model(model: Type[Model], enforce_lowercase: 
             )
         )
 
-    return errors
+    return messages
 
 
 DEFAULT_TABLE_NAME_CHECK_IGNORE = [
@@ -311,7 +349,7 @@ class CheckExplicitTableNames:
     ignore_labels: Iterable[str]
     enforce_lowercase: bool
 
-    def __init__(self, ignore_labels: Iterable[str] = DEFAULT_TABLE_NAME_CHECK_IGNORE, enforce_lowercase: bool = True):
+    def __init__(self, ignore_labels: Iterable[Union[str, Pattern]] = DEFAULT_TABLE_NAME_CHECK_IGNORE, enforce_lowercase: bool = True):
         """
         ignore_labels: ignore apps or models matching supplied labels
         enforce_lowercase: applies rule E010 which enforces table name to be all lowercase; defaults to True
@@ -319,15 +357,16 @@ class CheckExplicitTableNames:
         self.ignore_labels = ignore_labels
         self.enforce_lowercase = enforce_lowercase
 
-    def __call__(self, app_configs: Iterable[AppConfig], **kwargs):
+    def __call__(self, app_configs: Iterable[AppConfig], **kwargs) -> List[CheckMessage]:
         """
         Warn when models don't have Meta's db_table_name set in apps that require it.
         """
         candidate_models = find_candidate_models(app_configs, self.ignore_labels)
-        errors = []
+
+        messages = []
         for model in candidate_models.values():
-            errors += _check_explicit_table_names_on_model(model, self.enforce_lowercase)
-        return errors
+            messages += _check_explicit_table_names_on_model(model, self.enforce_lowercase)
+        return messages
 
 
 class CheckReversibleFieldNames:
@@ -343,7 +382,7 @@ class CheckReversibleFieldNames:
             errors += self._check_field_names_on_model(model)
         return errors
 
-    def _check_field_names_on_model(self, model: Type[Model]) -> Iterable[Type[Error]]:
+    def _check_field_names_on_model(self, model: Type[Model]) -> List[CheckMessage]:
         """
         check whether field names on model are legit
 
@@ -353,20 +392,20 @@ class CheckReversibleFieldNames:
 
         """
 
-        errors = []
+        messages = []
 
         for field in model._meta.fields:
             if camel_to_underscore(underscore_to_camel(field.name)) != field.name:
                 hint = None
                 if re.search(r'_[0-9]', field.name):
                     hint = f'Underscore before a number in {model._meta.label}.{field.name}'
-                errors.append(
+                messages.append(
                     Error(
-                        f"Field name is not reversible with underscore_to_camel()/camel_to_underscore()",
+                        "Field name is not reversible with underscore_to_camel()/camel_to_underscore()",
                         hint=hint,
                         obj=model,
                         id=ID_ERROR_FIELD_NAME_NOT_CAMEL_FRIENDLY,
                     )
                 )
 
-        return errors
+        return messages