+++ test_allianceutils/tests/isort_default_app_config/models.py
@@ -1,0 +1,2 @@
+# this is just a placeholder django app so that we can test isort config is being
+# loaded correctly (we override this first-party app into a third party app)