+++ test_allianceutils/tests/middleware/urls.py
@@ -1,4 +1,4 @@
-from django.conf.urls import url
+from django.urls import path
 
 from .views import current_user
 from .views import query_overhead
@@ -7,9 +7,7 @@ from .views import run_queries
 app_name = 'middleware'
 
 urlpatterns = [
-    url(r'^run_queries/$', run_queries, name='run_queries'),
-    url(r'^query_overhead/$', query_overhead, name='query_overhead'),
-
-    url(r'^current_user/$', current_user, name='current_user'),
-
+    path('run_queries/', run_queries, name='run_queries'),
+    path('query_overhead/', query_overhead, name='query_overhead'),
+    path('current_user/', current_user, name='current_user'),
 ]